+++ lib/sift/type_validator.rb
@@ -4,7 +4,6 @@ module Sift
     DATETIME_RANGE_PATTERN = { format: { with: /\A.+(?:[^.]\.\.\.[^.]).+\z/, message: "must be a range" }, valid_date_range: true }.freeze
     DECIMAL_PATTERN = { numericality: true, allow_nil: true }.freeze
     BOOLEAN_PATTERN = { inclusion: { in: [true, false] }, allow_nil: true }.freeze
-    JSON_PATTERN = { valid_json: true }.freeze
 
     WHITELIST_TYPES = [:int,
                        :decimal,
@@ -17,12 +16,13 @@ module Sift
                        :scope,
                        :jsonb].freeze
 
-    def initialize(param, type)
+    def initialize(param, type, keys = nil)
       @param = param
       @type = type
+      @keys = keys
     end
 
-    attr_reader :param, :type
+    attr_reader :param, :type, :keys
 
     def validate
       case type
@@ -35,7 +35,7 @@ module Sift
       when :boolean
         BOOLEAN_PATTERN
       when :jsonb
-        JSON_PATTERN
+        { valid_json: true, valid_jsonb: { keys: keys } }
       end
     end